fix save mail session

+++ src/controllers/defaults/session-basics.js
@@ -71,35 +71,49 @@ module.exports = {
 	async create(req, res, model, options) {
 		delete req.body.id
 		const permission = await Permissions.check(req.token, model.tableName, 'insert')
+		req.body.logist_id = await Sessions.getSessionId(req)
+		const checkLogist = await model.findOne({
+			where: { logist_id: req.body.logist_id },
+			include: [
+				{
+					model: Session,
+					as: 'sessions',
+					where: { mail: req.body.mail },
+				},
+			],
+		})
 		if (permission) {
-			req.body.logist_id = await Sessions.getSessionId(req)
-			req.body.password = await Server.getHash(req.body.password)
-			req.body.table = model.tableName
-			req.body.active = true
-			if (!isuuid(req.body.logist_id)) {
-				res.send({ status: 'LOGIST_IS_EMPTY' })
-			} else {
-				Sessions.create(req, (result) => {
-					if (result.status == 'SESSION_INSERT_SUCCESS') {
-						req.body.session_id = result.data.id
-						model
-							.build(req.body)
-							.save()
-							.then(async (data) => {
-								res.send({ status: model.tableName.toUpperCase() + '_INSERT_SUCCESS', data })
-							})
-							.catch(async (error) => {
-								await Session.destroy({ where: { id: req.body.session_id } })
-								res.send({
-									status: model.tableName.toUpperCase() + '_INSERT_ERROR',
-									error: error.parent ? error.parent.detail : error,
+			if (!checkLogist) {
+				req.body.password = await Server.getHash(req.body.password)
+				req.body.table = model.tableName
+				req.body.active = true
+				if (!isuuid(req.body.logist_id)) {
+					res.send({ status: 'LOGIST_IS_EMPTY' })
+				} else {
+					Sessions.create(req, (result) => {
+						if (result.status == 'SESSION_INSERT_SUCCESS') {
+							req.body.session_id = result.data.id
+							model
+								.build(req.body)
+								.save()
+								.then(async (data) => {
+									res.send({ status: model.tableName.toUpperCase() + '_INSERT_SUCCESS', data })
 								})
-								return
-							})
-					} else {
-						res.send({ status: 'SESSION_INSERT_ERROR', error: result.error })
-					}
-				})
+								.catch(async (error) => {
+									await Session.destroy({ where: { id: req.body.session_id } })
+									res.send({
+										status: model.tableName.toUpperCase() + '_INSERT_ERROR',
+										error: error.parent ? error.parent.detail : error,
+									})
+									return
+								})
+						} else {
+							res.send({ status: 'SESSION_INSERT_ERROR', error: result.error })
+						}
+					})
+				}
+			} else {
+				res.send({ status: 'MAIL_EXISTS', error: 'Email already exists in the system' })
 			}
 		} else {
 			res.send({ status: model.tableName.toUpperCase() + '_PERMISSION_ERROR', error: 'Action not allowed' })
